Added regression test for WHAM geometry based on Mana's PoP paper draft. This helped fix a subtle bug in the geometry calculator, where a function parameter was being cast (incorrectly) to a int. The generated grid looks good.

+++ zero/gkgeom.c
@@ -28,7 +28,7 @@ struct gkyl_gkgeom {
 
 // some helper functions
 static inline double
-choose_closest(int ref, double R[2], double out[2])
+choose_closest(double ref, double R[2], double out[2])
 {
   return fabs(R[0]-ref) < fabs(R[1]-ref) ? out[0] : out[1];
 }
